modules: dispatch handleEvent on the extended event id
EModuleEventType is the union of utils::eventType and LUNA_KEY_TYPE_E, but
switching on it converts to LUNA_KEY_TYPE_E alone, so every case label from
the base enum drew a -Wswitch warning. Switch on the int value instead: the
handlers genuinely dispatch over the whole extended range.

+++ src/modules/audioEffectManager/audioEffectManager.cpp
@@ -487,7 +487,7 @@ bool AudioEffectManager::_setAudioEqualizerPreset(LSHandle *lshandle, LSMessage 
 
 void AudioEffectManager::handleEvent(events::EVENTS_T *event)
 {
-    switch(event->eventName)
+    switch((int)event->eventName)
     {
         case  utils::eEventDeviceConnectionStatus:
         {
